refactored controller to its own file
