+++ videocall-client/src/diagnostics/mod.rs
@@ -1,0 +1,6 @@
+pub mod diagnostics;
+pub mod encoder_control_sender;
+
+// Re-export the modules
+pub use encoder_control_sender::{EncoderControl, EncoderControlSender};
+pub use diagnostics::{DiagnosticEvent, DiagnosticManager, SenderDiagnosticManager};
